+++ crates/iroh-http-core/src/session.rs
@@ -38,7 +38,7 @@ fn get_conn(endpoint: &IrohEndpoint, handle: u64) -> Result<Connection, CoreErro
         .handles()
         .lookup_session(handle)
         .map(|s| s.conn.clone())
-        .ok_or_else(|| CoreError::invalid_handle(handle as u32))
+        .ok_or_else(|| CoreError::invalid_handle(handle))
 }
 
 /// Return the remote peer's public key for a session.
@@ -155,7 +155,7 @@ pub fn session_close(
     let entry = endpoint
         .handles()
         .remove_session(session_handle)
-        .ok_or_else(|| CoreError::invalid_handle(session_handle as u32))?;
+        .ok_or_else(|| CoreError::invalid_handle(session_handle))?;
     let code = iroh::endpoint::VarInt::from_u64(close_code).map_err(|_| {
         CoreError::invalid_input(format!(
             "close_code {close_code} exceeds QUIC VarInt max (2^62 - 1)"
